+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### ♻️ Refactoring
 
+* [#95](https://github.com/antoinezanardi/werewolves-assistant-web/issues/95) - Change game options structure.
 * [#105](https://github.com/antoinezanardi/werewolves-assistant-web/issues/105) - Change `attributes.attribute` to `attributes.name` in Player class.
 
 ### 📦 Packages